+++ lib/pdfmd/sort.rb
@@ -6,26 +6,34 @@
 inputDir = ENV.fetch('PDFMD_INPUTDIR')
 
 require_relative('./methods.rb')
+require_relative '../string_extend.rb'
 require 'fileutils'
 
 opt_destination = ENV.fetch('PDFMD_DESTINATION')
 opt_dryrun      = ENV.fetch('PDFMD_DRYRUN') == 'true' ? true : false
 opt_copy        = ENV.fetch('PDFMD_COPY')
 opt_log         = ENV.fetch('PDFMD_LOG')
+opt_logfilepath = ENV.fetch('PDFMD_LOGFILEPATH')
 opt_interactive = ENV.fetch('PDFMD_INTERACTIVE')
+hieraDefaults   = queryHiera('pdfmd::config')
 
-hieraDefaults = queryHiera('pdfmd::config')
-
-copyAction  = opt_copy.empty? ? false : true
-if opt_copy.nil? and hieraDefaults['sort']['copy'] == true
+# Determin the setting for the copy/move action when sorting
+# Use HieraDefaults if nothing has been set.
+copyAction  = opt_copy
+if copyAction.blank? and hieraDefaults['sort']['copy'] == true
   copyAction = true
-  puts 'Setting action to copy based on Hiera.'
+elsif copyAction.blank? or copyAction == 'false'
+  copyAction = false
 end
 
-interactiveAction = opt_interactive.empty? ? false : true
-if opt_interactive.empty? and hieraDefaults['sort']['interactive'] == true
+# Determine the setting for interaction
+if opt_interactive.blank? and hieraDefaults['sort']['interactive'] == true
+  puts 'Setting interactive from hiera'
   interactiveAction = true
-  puts 'Setting interactive to true based on Hiera.'
+elsif opt_interactive == 'true'
+  interactiveAction = true
+elsif opt_interactive.blank? or opt_interactive == 'false'
+  interactiveAction = false
 end
 
 # Fetch alternate destination from hiera if available
@@ -41,18 +49,47 @@ if destination.nil? or destination == ''
     puts 'Abort.'
     exit 1
   end
-
 end
 
-logenable = opt_log
-logfile   = !hieraDefaults['sort']['logfile'].nil? ? hieraDefaults['sort']['logfile'] : Dir.pwd.chomp('/') + '/' + Pathname.new(__FILE__).basename + '.log'
-
-# Check that logfilepath exists and is writeable
-if !File.writable?(logfile)
-  puts "Cannot write '#{logfile}. Abort."
-  exit 1
+# Determine the state of the logging
+if (opt_log.blank? and hieraDefaults['sort']['log'] == true) or
+  opt_log == 'true'
+  logenable = true
+elsif opt_log.blank? or opt_log == 'false'
+  logenable = false
 end
-logenable ? $logger = Logger.new(logfile) : ''
+
+if logenable
+
+  if opt_logfilepath.blank? and 
+    ( hieraDefaults['sort']['logfilepath'].nil? or
+     hieraDefaults['sort']['logfilepath'].blank? or
+     hieraDefaults['sort'].nil? )
+
+    logfile = Dir.pwd.chomp('/') + '/' + File.basename(ENV['PDFMD'], '.*') + '.log'
+
+  elsif not opt_logfilepath.blank?
+
+    if File.directory? opt_logfilepath
+      abort('Logfilepath is a directory. Abort.')
+      exit 1
+    end
+
+    logfile = opt_logfilepath
+
+  elsif opt_logfilepath.blank? and
+    not hieraDefaults['sort']['logfilepath'].blank? 
+
+    logfile = hieraDefaults['sort']['logfilepath']
+
+  else
+
+    logfile = Dir.pwd.chomp('/') + '/' + File.basename(ENV['PDFMD'], '.*') + '.log'
+
+  end
+
+  $logger = Logger.new(logfile)
+end
 
 # Input validation
 !File.exist?(inputDir) ? abort('Input directory does not exist. Abort.'): ''
@@ -60,7 +97,7 @@ File.directory?(inputDir) ? '' : abort('Input is a single file. Not implemented 
 File.file?(destination) ? abort("Output '#{destination}' is an existing file. Cannot create directory with the same name. Abort") : ''
 unless File.directory?(destination)
   FileUtils.mkdir_p(destination)
-  $logger.info("Destination '#{destination}' has been created.")
+  logenable ? $logger.info("Destination '#{destination}' has been created.") : ''
 end
 
 # Iterate through all files
@@ -69,6 +106,7 @@ Dir[inputDir.chomp('/') +  '/*.pdf'].sort.each do |file|
   if interactiveAction
     answer = readUserInput("Process '#{file}' ([y]/n): ")
     answer = answer.empty? ? 'y' : answer 
+    logenable ? $logger.info("Interactive answer for file '#{file}' : #{answer}") : ''
     answer.match(/y/) ? '' : next
   end
 
@@ -89,13 +127,14 @@ Dir[inputDir.chomp('/') +  '/*.pdf'].sort.each do |file|
 
     # Final check before touching the filesystem
     if not File.exist?(filedestination)
-      $logger.info("File '#{file}' => '#{filedestination}'")
 
       # Move/Copy the file
-      if copyAction and not opt_dryrun
-        #FileUtils.cp(file, filedestination)
-      elsif not opt_dryrun
-        #FileUtils.mv(file,filedestination)
+      if copyAction 
+        opt_dryrun ? '' : FileUtils.cp(file, filedestination)
+        logenable ? $logger.info("File copied '#{file}' => '#{filedestination}'") : ''
+      else
+        opt_dryrun ? '' : FileUtils.mv(file,filedestination)
+        logenable ? $logger.info("File moved '#{file}' => '#{filedestination}'") : ''
       end
 
     else